+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - CMS-1640 - Move ISD pages to great international
 - CI-383 - Create About DIT landing page
 - CI-353 - Invest International Home page uses directory components card and wagtails focal area on images
+- CMS-1745 - Implement invest redirects
 - CI-420 - Added redirect from Capital Invest url ('/international/invest-capital') to one using tree based routing (/international/content/capital-invest')